An earlier vimrc
This is not my most basic, as that is lost to time, but it is about the
simplest that I've got. It unmaps the pesky arrow keys, adds a convenient
mapping to edit my vimrc (and autoload it once saved), and sets my preferred
indentation, with the exception for Makefiles.  Also enables wildmenu and
hidden files.